Bootanimation.zip
Created a bootanimation for GrapheneOS.
This needs testing as I followed (outdated?) instructions from here https://forum.xda-developers.com/showthread.php?t=1096068. I was unable to get this working - could push bootanimation to test device but could not verify it on the actual device. Not sure if instructions are for an older version of Android. 
I can provide any source file for the animation as well if anyone else wants to give this a try.